Disable old non-functioning processors and tests
The NullDereference processor is not functioning. Neither its old tests not the added ones from the SonarQube rules documentation are fixed by the processor. To fix, the processor needs a complete rewrite.
The NonSerializableSuperClass processor may be working but its test case is not. The test cannot find anything wrong with the initial state even though the example (which should be failing) is taken directly from the SonarQube rules definition. I've been unable to understand why or fix it. Seeing as we have no test verifying the processor, disabling it is safest.