WIP Reimplementing sniper mode
Sniper mode has been updated and the old implementation in sonarqube-repair
was non-functional. This is the first step to fixing it.

It works to an extent; 2/4 tested files work fine but 2 fail with error message [num substituted for actual numbers]:
spoon.SpoonException: Cannot compare this: [num, num] with other: ["num", "num"].
The two failing files are included in the commit, AbstractSourceFragmentContext.java & CodeFactory.java.
Both are taken directly from the spoon-core repository.

For Codefactory, the offending line is line 200 (CtConstructor<T> constructor =...)
while for AbstractSrc, 4 methods are breaking. They are:
    findIndexOfNextChildTokenByValue
    findIndexOfNextChildTokenByType
    findIndexOfNextChildTokenOfRole
    findIndexOfNextChildTokenOfElement