WIP:Failing old processor and test case
This is an old implementation of a processor for Sonarqube rule S2055 with a failing test case.
It is recommended that an attempt at fixing it changes the filter to use Spoon rather than the Sonarjava plugin.
For examples of this, look at other processors of type BUG (except S2095).